Remove top-level `_utils`
