+++ src/moscot/plotting/_utils.py
@@ -1,3 +1,4 @@
+import contextlib
 from collections import defaultdict
 from copy import copy
 from types import MappingProxyType
@@ -9,6 +10,7 @@ from typing import (
     Mapping,
     Optional,
     Sequence,
+    Set,
     Tuple,
     Union,
 )
@@ -34,8 +36,6 @@ from scanpy.plotting._utils import (
 
 if TYPE_CHECKING:
     from moscot.base.problems.compound_problem import CompoundProblem
-
-from moscot._utils import RandomKeys
 
 
 def set_palette(
@@ -553,3 +553,53 @@ def get_plotting_vars(adata: AnnData, func_key: str, *, key: str) -> Any:
         return adata.uns[uns_key][func_key][key]
     except KeyError:
         raise KeyError(f"No data found in `adata.uns[{uns_key!r}][{func_key!r}][{key!r}]`.") from None
+
+
+class RandomKeys:
+    """
+    Create random keys inside an :class:`~anndata.AnnData` object.
+
+    Parameters
+    ----------
+    adata
+        Annotated data object.
+    n
+        Number of keys, If `None`, create just 1 keys.
+    where
+        Attribute of ``adata``. If `'obs'`, also clean up `'{key}_colors'` for each generated key.
+
+    """
+
+    def __init__(self, adata: AnnData, n: Optional[int] = None, where: str = "obs"):
+        self._adata = adata
+        self._where = where
+        self._n = n or 1
+        self._keys: List[str] = []
+
+    def _generate_random_keys(self):
+        def generator() -> str:
+            return f"RNG_COL_{np.random.RandomState().randint(2 ** 16)}"
+
+        where = getattr(self._adata, self._where)
+        names: List[str] = []
+        seen: Set[str] = set(where.keys())
+
+        while len(names) != self._n:
+            name = generator()
+            if name not in seen:
+                seen.add(name)
+                names.append(name)
+
+        return names
+
+    def __enter__(self):
+        self._keys = self._generate_random_keys()
+        return self._keys
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        with contextlib.suppress(KeyError):
+            for key in self._keys:
+                df = getattr(self._adata, self._where)
+                df.drop(key, axis="columns", inplace=True)
+                if self._where == "obs":
+                    del self._adata.uns[f"{key}_colors"]
